save.c up to take_bonus

+++ include/collision/bit_util.h
@@ -10,6 +10,6 @@ s16 cosinus(s16 param_1);
 s16 sinus(s16 param_1);
 
 s32 ashl16(s16 param_1, u32 param_2); /* TODO: is this related with collision at all? */
-s16 ashr16(u16 param_1, u32 param_2);
+s32 ashr16(u16 param_1, u32 param_2);
 
 #endif
